Update Program.cs
Added Regex and openSpots, removed some unnecessary code. This new added code controls what to do if a user doesn't enter an integer as well as handling when a random number is generated and it happens to be a spot that has already been chosen. Note to self: When while loop has exited, if something happens that meets the condition it enters the loop again without being told to run the code the loop is within. Specifically gameLoop ( ) only needs to be called once. It is still being called until the entire loop meets the condition to exit.